Update gulpfile to pull all js files from js dir. Move multilanguage js to its own file
